feat: quiz in classes
