+++ lms/templates/quiz/quiz.html
@@ -118,9 +118,6 @@
 			<div class="btn btn-secondary btn-sm hide" id="summary">
 				{{ _("Submit") }}
 			</div>
-			<small id="submission-message" class="font-weight-bold hide">
-				{{ _("Please join the course to submit the Quiz.") }}
-			</small>
 			<div class="btn btn-secondary btn-sm hide" id="try-again">
 				{{ _("Try Again") }}
 			</div>
