+++ lms/www/quiz_submission/quiz_submission.html
@@ -1,0 +1,58 @@
+{% extends "lms/templates/lms_base.html" %}
+{% block title %}
+    {{ quiz.title }}
+{% endblock %}
+
+{% block page_content %}
+<main class="common-page-style">
+    {{ Header() }}
+    <div class="container form-width">
+        {{ SubmissionForm(quiz) }}
+    </div>
+</main>
+{% endblock %}
+
+{% macro Header() %}
+<header class="sticky mb-5">
+    <div class="container form-width">
+        <div class="edit-header">
+            <div>
+                <div class="vertically-center">
+                    <div class="page-title">
+                        {{ quiz.title }}
+                    </div>
+                    {% if submission.score %}
+                    <div>
+                        {{ submission.score }}
+                    </div>
+                    {% endif %}
+                </div>
+                <div class="vertically-center small">
+                    <a class="dark-links" href="/classes">
+                        {{ _("All Classes") }}
+                    </a>
+                </div>
+            </div>
+
+
+            <div class="align-self-center">
+                <!-- <button class="btn btn-primary btn-sm btn-save-assignment" {% if quiz.name %} data-quiz="{{ quiz.name }}" {% endif %}
+                    {% if submission.name %} data-submission="{{ submission.name }}" {% endif %}>
+                    {{ _("Save") }}
+                </button> -->
+            </div>
+        </div>
+    </div>
+</header>
+{% endmacro %}
+
+{% macro SubmissionForm(quiz) %}
+    {% include("lms/templates/quiz/quiz.html") %}
+{% endmacro %}
+
+{% block script %}
+    {{ super() }}
+    <script type="text/javascript">
+        {% include "lms/templates/quiz/quiz.js" %}
+    </script>
+{% endblock %}
